Extract class RealtyScene
